refactor(router): split remainder of describes in integration test file
to more easily navigate and identify different suites of integration tests,
this commit splits them into several different files
